// src/app/Interfaces/persona.interface.ts
export interface personaI {
    id?:Number;
    //usuario: String;
    nombre: String;
    apellido: String;
    //password: String;
    ocupacion: String;
    acercaDe: String;
    //fechaNac: String;
    //nacionalidad: String;
    mail: String;
    //sobreMi: String;
    urlImgBackground: String;
    urlImgPerfil: String;
}